# backend/src/main/java/com/medtrack/auth/controller/AuthController.java
package com.medtrack.auth.controller; import com.medtrack.auth.dto.*; import com.medtrack.auth.service.AuthService; import jakarta.servlet.http.HttpServletRequest; import jakarta.validation.Valid; import org.springframework.http.*; import org.springframework.web.bind.annotation.*;
@RestController @RequestMapping("/api/v1/auth") public class AuthController {private static final String COOKIE="medtrack_refresh";private final AuthService service;public AuthController(AuthService s){service=s;}@PostMapping("/login")public ResponseEntity<AuthResponse> login(@Valid @RequestBody LoginRequest r){return respond(service.login(r));}@PostMapping("/refresh")public ResponseEntity<AuthResponse> refresh(@CookieValue(name=COOKIE,required=false)String token){if(token==null)throw new com.medtrack.shared.exception.DomainException("MISSING_REFRESH_TOKEN","Refresh token cookie is required");return respond(service.refresh(token));}@PostMapping("/logout")public ResponseEntity<Void> logout(@CookieValue(name=COOKIE,required=false)String token){if(token!=null)service.logout(token);return ResponseEntity.noContent().header(HttpHeaders.SET_COOKIE,ResponseCookie.from(COOKIE,"").httpOnly(true).sameSite("Strict").path("/api/v1/auth").maxAge(0).build().toString()).build();}private ResponseEntity<AuthResponse> respond(AuthService.AuthResult r){ResponseCookie c=ResponseCookie.from(COOKIE,r.refreshToken()).httpOnly(true).sameSite("Strict").secure(false).path("/api/v1/auth").maxAge(7*86400).build();return ResponseEntity.ok().header(HttpHeaders.SET_COOKIE,c.toString()).body(r.response());}}