# backend/src/main/java/com/medtrack/inventory/controller/InventoryController.java
package com.medtrack.inventory.controller;
import com.medtrack.inventory.dto.*; import com.medtrack.inventory.repository.InventoryBalanceRepository; import com.medtrack.inventory.service.*; import jakarta.validation.Valid; import java.util.*; import org.springframework.data.domain.*; import org.springframework.http.*; import org.springframework.security.access.prepost.PreAuthorize; import org.springframework.security.core.annotation.AuthenticationPrincipal; import org.springframework.web.bind.annotation.*;
@RestController @RequestMapping("/api/v1/inventory") public class InventoryController {private final InboundService inbound;private final FefoAllocationEngine fefo;private final InventoryBalanceRepository balances;public InventoryController(InboundService i,FefoAllocationEngine f,InventoryBalanceRepository b){inbound=i;fefo=f;balances=b;}@PostMapping("/inbound")@PreAuthorize("hasRole('SUPER_ADMIN') or hasRole('CENTRAL_WAREHOUSE_MANAGER')")public ResponseEntity<InboundReceiptResponse> inbound(@AuthenticationPrincipal String actorId,@RequestHeader(name="X-Idempotency-Key",required=false)String key,@Valid @RequestBody InboundReceiptRequest request){InboundReceiptResponse response=inbound.receive(UUID.fromString(actorId),key,request);return ResponseEntity.status(HttpStatus.CREATED).body(response);}@GetMapping("/balances")@PreAuthorize("hasRole('SUPER_ADMIN') or hasRole('CENTRAL_WAREHOUSE_MANAGER') or hasRole('STORE_MANAGER') or hasRole('AUDITOR')")public Page<InventoryBalanceResponse> balances(@RequestParam UUID warehouseId,@RequestParam(defaultValue="0")int page,@RequestParam(defaultValue="20")int size){return balances.findByWarehouse_Id(warehouseId,PageRequest.of(page,Math.min(Math.max(size,1),100))).map(InventoryBalanceResponse::of);}@PostMapping("/allocations/fefo")@PreAuthorize("hasRole('SUPER_ADMIN') or hasRole('CENTRAL_WAREHOUSE_MANAGER')")public FefoAllocationResponse allocate(@AuthenticationPrincipal String actorId,@Valid @RequestBody FefoAllocationRequest request){return fefo.allocate(UUID.fromString(actorId),request);}@PostMapping("/allocations/manual")@PreAuthorize("hasRole('SUPER_ADMIN') or hasRole('CENTRAL_WAREHOUSE_MANAGER')")public FefoAllocationResponse override(@AuthenticationPrincipal String actorId,@Valid @RequestBody FefoOverrideRequest request){return fefo.override(UUID.fromString(actorId),request);}}