# backend/src/main/java/com/medtrack/medicine/controller/MedicineController.java
package com.medtrack.medicine.controller; import com.medtrack.medicine.dto.*; import com.medtrack.medicine.service.MedicineService; import jakarta.validation.Valid; import java.util.*; import org.springframework.data.domain.*; import org.springframework.http.*; import org.springframework.security.access.prepost.PreAuthorize; import org.springframework.web.bind.annotation.*;
@RestController @RequestMapping("/api/v1/medicines") public class MedicineController {private final MedicineService service; public MedicineController(MedicineService s){service=s;} @GetMapping @PreAuthorize("hasAuthority('MEDICINE_READ')") public Page<MedicineResponse> list(@RequestParam(required=false)String search,@RequestParam(defaultValue="0")int page,@RequestParam(defaultValue="20")int size,@RequestParam(defaultValue="sku")String sort){return service.list(search,PageRequest.of(page,Math.min(Math.max(size,1),100),Sort.by(sort)));} @PostMapping @PreAuthorize("hasAuthority('MEDICINE_CREATE')") public ResponseEntity<MedicineResponse> create(@Valid @RequestBody MedicineRequest r){return ResponseEntity.status(HttpStatus.CREATED).body(service.create(r));} @GetMapping("/{id}") @PreAuthorize("hasAuthority('MEDICINE_READ')") public MedicineResponse get(@PathVariable UUID id){return service.get(id);} @PutMapping("/{id}") @PreAuthorize("hasAuthority('MEDICINE_UPDATE')") public MedicineResponse update(@PathVariable UUID id,@Valid @RequestBody MedicineRequest r){return service.update(id,r);} }